Moves ByteSizeHelper to more approiate location
